requirements.txt
aioredis==1.3.1
apturl==0.5.2
arabic-reshaper==2.1.3
asgi-redis==1.4.3
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.2
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==36.0.2
cssselect2==0.5.0
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.13
django-allauth==0.50.0
django-encrypted-id==0.3.2
django-extensions==3.1.5
django-phonenumber-field==6.1.0
django-safedelete==1.1.2
django-sslserver==0.22
django-twilio==0.13.2
djangorestframework==3.13.1
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Fiona==1.8.21
fpdf==1.7.2
future==0.18.2
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
gpg===1.13.1-unknown
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
monotonic==1.5
msgpack==1.0.3
msgpack-python==0.5.6
munch==2.5.0
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
oscrypto==1.3.0
packaging==20.3
pandas==1.4.2
paramiko==2.6.0
pexpect==4.6.0
pgeocode==0.3.0
phonenumbers==8.12.46
Pillow==9.1.0
protobuf==3.6.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==2.4.6
PyPDF3==1.0.6
pyproj==3.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
qrcode==7.3.1
redis==2.10.6
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==21.1.0
Shapely==1.8.1.post1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
ssh-import-id==5.10
svglib==1.2.1
systemd-python==234
tinycss2==1.1.1
tqdm==4.64.0
twilio==6.63.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritools==4.0.0
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
webencodings==0.5.1
websockets==10.3
Werkzeug==2.1.2
xhtml2pdf==0.2.7
xkit==0.0.0
zope.interface==5.4.0
